Add basic support for multiple pub, subs
-Previously, multiple calls to advertise/subscribe
 would return the same publisher/subscriber so
 shutting down 1 would shutdown all of them.
-Also starts to add support for max number of
 attempts during xmlrpc calls.
