+++ src/utils/event_utils.js
@@ -1,0 +1,5 @@
+module.exports = {
+  rebroadcast(evt, emitter, rebroadcaster) {
+    emitter.on(evt, rebroadcaster.emit.bind(rebroadcaster, evt));
+  }
+};
